+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.7.5] - 2023-05-28
 ### Changed
 - RaceDetailView to RaceTahminView
+- Refactor NewTahminView
 
 ## [1.7.4] - 2023-05-27
 ### Fixed